requirements.txt
pdfplumber==0.11.0
google-generativeai==0.4.1
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
